Move user code into user
